backend/embedding/requirements.txt
sentence-transformers
numpy
grpcio
grpcio-tools
protobuf
python-dotenv
torch